remove "flattened" style of picked resource items
When a resource list item is picked, it changes it's style to a single
border line. This didn't look good when a large number of items were
picked, and didn't really indicate anything useful.

This commit removes that style completely. When an item is picked from
the list, it leaves nothing visible.